refactor(store-postgres): share active logical graph locking
