+++ crates/automata-ci-store-postgres/src/logical_activation_preparation.rs
@@ -9,7 +9,8 @@ use sqlx::{Postgres, Row as _, Transaction, postgres::PgRow};
 use uuid::Uuid;
 
 use super::{
-    PostgresStore, durable_schema::current_durable_schemas, pg_bigint,
+    PostgresStore, durable_schema::current_durable_schemas,
+    logical_graph::lock_active_logical_graph, pg_bigint,
     workflow_runtime_policy::load_pinned_runtime_policy_for_run,
 };
 use automata_ci_store::{
@@ -76,7 +77,14 @@ impl LogicalActivationPreparationStore for PostgresStore {
             return Err(LogicalActivationPreparationStoreError::ClaimRejected);
         }
         lock_preparation_quarantine_custody(&mut transaction, request.claim()).await?;
-        if !lock_active_preparation_graph(&mut transaction, request.claim().target()).await? {
+        if !lock_active_logical_graph(
+            &mut transaction,
+            request.claim().target().tenant(),
+            request.claim().target().run_id(),
+            request.claim().target().invocation_id(),
+        )
+        .await?
+        {
             return Err(LogicalActivationPreparationStoreError::ClaimRejected);
         }
         let row = lock_target(&mut transaction, request.claim().target())
@@ -167,7 +175,14 @@ impl LogicalActivationPreparationStore for PostgresStore {
     ) -> Result<LogicalActivationPreparationReceipt, LogicalActivationPreparationStoreError> {
         let mut transaction = self.pool.begin().await.map_err(operation_error)?;
         lock_preparation_continuation_custody(&mut transaction, request.claim()).await?;
-        if !lock_active_preparation_graph(&mut transaction, request.claim().target()).await? {
+        if !lock_active_logical_graph(
+            &mut transaction,
+            request.claim().target().tenant(),
+            request.claim().target().run_id(),
+            request.claim().target().invocation_id(),
+        )
+        .await?
+        {
             return Err(LogicalActivationPreparationStoreError::ClaimRejected);
         }
         let row = lock_target(&mut transaction, request.claim().target())
@@ -477,7 +492,14 @@ pub(super) async fn claim_logical_activation_preparation_in_transaction(
     request: &ClaimLogicalActivationPreparation,
     origin_selection_id: Uuid,
 ) -> Result<LogicalActivationPreparationClaimOutcome, LogicalActivationPreparationStoreError> {
-    if !lock_active_preparation_graph(transaction, request.target()).await? {
+    if !lock_active_logical_graph(
+        transaction,
+        request.target().tenant(),
+        request.target().run_id(),
+        request.target().invocation_id(),
+    )
+    .await?
+    {
         return Ok(LogicalActivationPreparationClaimOutcome::NotReady);
     }
     let row = lock_target(transaction, request.target())
@@ -801,71 +823,6 @@ async fn lock_preparation_continuation_custody(
 ) -> Result<(), LogicalActivationPreparationStoreError> {
     lock_preparation_selection_custody(transaction, claim).await?;
     lock_preparation_quarantine_custody(transaction, claim).await
-}
-
-async fn lock_active_preparation_graph(
-    transaction: &mut Transaction<'_, Postgres>,
-    target: &LogicalActivationPreparationTarget,
-) -> Result<bool, LogicalActivationPreparationStoreError> {
-    let schemas = current_durable_schemas();
-    let run_active: Option<bool> = sqlx::query_scalar(
-        r"
-        SELECT run.status IN ('queued', 'in_progress')
-               AND run.admission_epoch = $4 AND run.plan_schema = $3
-        FROM workflow_runs AS run
-        JOIN repositories AS repository ON repository.id = run.repository_id
-        WHERE repository.tenant_id = $1 AND run.id = $2
-        FOR SHARE OF run
-        ",
-    )
-    .bind(target.tenant().as_str())
-    .bind(target.run_id().as_uuid())
-    .bind(schemas.workflow_plan_i32)
-    .bind(schemas.admission_epoch_i32)
-    .fetch_optional(&mut **transaction)
-    .await
-    .map_err(operation_error)?;
-    if run_active != Some(true) {
-        return Ok(false);
-    }
-    let marker_active: Option<bool> = sqlx::query_scalar(
-        r"
-        SELECT marker.state IN ('pending', 'active')
-               AND marker.orchestration_schema = $3
-               AND marker.admission_graph_sealed_at_ms IS NOT NULL
-               AND automata_logical_workflow_invocation_published(
-                   marker.run_id, $2
-               )
-        FROM logical_workflow_runs AS marker
-        WHERE marker.run_id = $1
-        FOR SHARE OF marker
-        ",
-    )
-    .bind(target.run_id().as_uuid())
-    .bind(target.invocation_id().as_uuid())
-    .bind(schemas.logical_orchestration_i16)
-    .fetch_optional(&mut **transaction)
-    .await
-    .map_err(operation_error)?;
-    if marker_active != Some(true) {
-        return Ok(false);
-    }
-    let invocation_active: Option<bool> = sqlx::query_scalar(
-        r"
-        SELECT invocation.state IN ('pending', 'active')
-               AND invocation.plan_schema = $3
-        FROM logical_workflow_invocations AS invocation
-        WHERE invocation.run_id = $1 AND invocation.id = $2
-        FOR SHARE OF invocation
-        ",
-    )
-    .bind(target.run_id().as_uuid())
-    .bind(target.invocation_id().as_uuid())
-    .bind(schemas.workflow_plan_i16)
-    .fetch_optional(&mut **transaction)
-    .await
-    .map_err(operation_error)?;
-    Ok(invocation_active == Some(true))
 }
 
 async fn reject_quarantined_preparation(
