+++ crates/automata-ci-store-postgres/src/schema_bindings_tests.rs
@@ -34,13 +34,14 @@ fn admission_epoch_and_workflow_plan_use_independent_sql_parameters() {
         ("github_job_runtime_authority.rs", 1),
         ("github_oidc.rs", 1),
         ("github_subject_evidence.rs", 2),
-        ("logical_activation.rs", 5),
-        ("logical_activation_preparation.rs", 2),
+        ("logical_activation.rs", 4),
+        ("logical_activation_preparation.rs", 1),
+        ("logical_graph.rs", 1),
         ("logical_instance_result.rs", 1),
         ("logical_job_result.rs", 1),
         ("logical_materialization.rs", 3),
         ("logical_run_finalization.rs", 2),
-        ("logical_work_selection.rs", 6),
+        ("logical_work_selection.rs", 5),
         ("managed_secret_authority.rs", 2),
         ("reusable_workflow_runtime.rs", 1),
     ] {
